# Improve performance of get_direction function